Create no console window for spawned `commands`
Spawned commands related to `adb` and `scrcpy` create a console
terminal when called from Tauri powered GUI apps in Windows. This
patch fixes that by setting the `CREATE_NO_WINDOW` flag when
spawning the commands in Windows.

This patch also includes fixes for clippy suggestions introduced
in the latest Rust stable (1.65).

In addition, this patch also changes the default `scrcpy` and `adb`
paths to search the current executable directory first (different
from working directory sometimes) and then the `PATH` environment
variable. This is because when bundled, the `adb` and `scrcpy`
binaries are placed in the same directory as the `adborc` executable.