Add version checking for connections
This patch adds support for version checking for supplier and
consumer connections. Currently (0.1.0), we only allow connections
from supplier/consumer that are of the same version as the
marketmaker. In the future, we may want to allow connections from
the same major version, but for now, we will only allow connections
from the same version as the marketmaker, until we have a stable
API (1.0 release).