Remove more redundant clones in `cli.rs`
This patch removes more redundant clones which were missed in
3c4c38f7021db0ee409eccc78f0d8def4027db58.

It also, changes `.clone()` to `.to_owned()` for mangen
as it is more descriptive.

Further, it moves the function `mangen()` to the bottom of the file
as it is not used for the core functionality of the program
and is only available for crate feature `mangen`.