Make it easier to add supported scrcpy arguments
This patch makes the changes required to add support for new scrcpy
arguments easier and contained to a small location. To add a new
argument, modify the `Scrcpy` and `SetScrcpyArgs` variants of the
`ConsumerCommands` enum in src/cli.rs. Also, modify the `ScrCpyArgs`
enum in src/util.rs to add the new argument. Additionally, might
need to modify `adb_utils::get_min_required_version` depending on
the new argument.

Note: This change breaks backwards compatibility with the previous
version of AdbOrc. That is not a problem since no stable API
has been released or promised yet and the project is still in alpha.
But that does mean that version numbers will be bumped to 0.2.0
and 0.1.0 nodes will no longer be compatible with 0.2.0 nodes.