Add a `scrcpy_utils` module
This patch further simplifies adding support for
new scrcpy arguments. This is done by refactoring all the
scrcpy arguments related code to its own module. This module
is not directly exported to the user of the library. But some
structs, enums and functions are made public by re-exporting
them from the `util::adb_utils` module.

Included with this change is the addition of support for a bunch
of new scrcpy arguments.