Make it easier to add supported scrcpy arguments (#1)
* Make it easier to add supported scrcpy arguments

This patch simplifies adding support for
new scrcpy arguments. This is done by refactoring all the
scrcpy arguments related code to its own module. This module
is not directly exported to the user of the library. But some
structs, enums and functions are made public by re-exporting
them from the `util::adb_utils` module.

This patch makes the changes required to add support for new scrcpy
arguments easier and contained to a small location. To add a new
argument, modify the `ScrcpyCliArgs` struct in the `scrcpy_utils` module.
Also, modify the `ScrCpyArgs` enum in `scrcpy_utils` to add the new
argument. Additionally, might need to modify 
`scrcpy_utils::get_min_required_version` depending on
the new argument.

Note: This change breaks backwards compatibility with the previous
version of AdbOrc. That is not a problem since no stable API
has been released or promised yet and the project is still in alpha.
But that does mean that version numbers will be bumped to 0.2.0
and 0.1.0 nodes will no longer be compatible with 0.2.0 nodes.

* Add `get-scrcpy-args` sub-command

Included with this change is the addition of support for a bunch
of new scrcpy arguments.