Add client version check (#8)
This patch partially resolves #4 by adding a version check to the
client. The client will now send a version check request to the
server and will not proceed if the server version is not compatible.

The reason this only partially resolves #4 is that the for client
version pre-0.2.0, this check is not performed and the client will
proceed without version check, possibly encountering an unhelpful
error message later on.