Contain usage of serde_json to `request` module
This patch contains usage of serde_json to `request` module. It
does this by introducing ToJson trait which is implemented for
all Response types. Also, implementation of std::str::FromStr
for `Request` and all response types is added.

In addition, this patch adds a `send_request` function to the `cli`
module which abstract a lot of the boilerplate code needed to
send a request and receive a response.

This is a mostly refactor patch, and it should not change any
functionality.