Remove `GetRequest` trait
This patch removes the `GetRequest` trait and abstracts everything
into `ToJson` trait. This is because the `GetRequest` trait is
only needed when we want to convert a request into a JSON string.
So the additional trait of wrapping the request into a `Request`
enum (using `GetRequest`) and then converting it into a JSON string
is redundant. Instead, a `ToJson` trait is implemented for each
request type which converts the request into a JSON string.