Limit usage of serde_json to `request` module (#14)
* Limit usage of serde_json to `request` module

This patch limits all usage of serde_json to `request` module. It
does this by introducing ToJson trait which is implemented for
all Response types. Also, implementation of std::str::FromStr
for `Request` and all response types is added.

In addition, this patch adds a `send_request` function to the `cli`
module which abstract a lot of the boilerplate code needed to
send a request and receive a response.

This is a mostly refactor patch, and it should not change any
functionality.


* TCPClent wrap request and send

This changes the TCPClent `send_request` and `send_no_wait` to
wrap the request in appropriate `Request` enum before sending.

* Remove `GetRequest` trait

This patch removes the `GetRequest` trait and abstracts everything
into `ToJson` trait. This is because the `GetRequest` trait is
only needed when we want to convert a request into a JSON string.
So the additional trait of wrapping the request into a `Request`
enum (using `GetRequest`) and then converting it into a JSON string
is redundant. Instead, a `ToJson` trait is implemented for each
request type which converts the request into a JSON string.

* Remove `ResponseType` from `cli.rs`

The changes made previously in this PR have rendered the
`ResponseType` enum in `cli.rs` obsolete. This commit removes it
and removes the `RequestProcessingError` variant for all response
types except, `SysStateResponse`.